\name{gen_x}
\alias{error_check}
\alias{gen_x}
\alias{initialize}
\alias{lower}
\alias{s_x}
\alias{sample}
\alias{sample_from_S}
\alias{show}
\alias{update}
\alias{upper}
\title{Cadapt_reject_sample generating first two points}
\arguments{
  \item{object}{\code{\linkS4class{Cadapt_reject_sample}}
  object}

  \item{object}{\code{\linkS4class{Cadapt_reject_sample}}
  object}

  \item{object}{\code{\linkS4class{Cadapt_reject_sample}}
  object}

  \item{object}{\code{\linkS4class{Cadapt_reject_sample}}
  object}

  \item{object}{\code{\linkS4class{Cadapt_reject_sample}}
  object}

  \item{object}{\code{\linkS4class{Cadapt_reject_sample}}
  object}

  \item{object}{\code{\linkS4class{Cadapt_reject_sample}}
  object}

  \item{object}{\code{\linkS4class{Cadapt_reject_sample}}
  object}

  \item{object}{\code{\linkS4class{Cadapt_reject_sample}}
  object}

  \item{object}{\code{\linkS4class{Cadapt_reject_sample}}
  object}
}
\description{
  Cadapt_reject_sample generating first two points

  Cadapt_reject_sample initialization

  Cadapt_reject_sample show

  Cadapt_reject_sample error_check

  Function to normalize the upper bounds of log(f(x))

  Cadapt_reject_sample upper

  Cadapt_reject_sample lower

  Cadapt_reject_sample sample

  Cadapt_reject_sample sample_from_S

  Cadapt_reject_sample update
}

