\name{ARSpackage-package}
\alias{ARSpackage-package}
\alias{ARSpackage}
\docType{package}
\title{
{ARSpackage:} an Adaptive Rejection Sampler
}
\description{
Final project for Statistics 243, an R package that performs adaptive rejection sampling, first proposed by Gilks and Wild in 1992.
}
\details{
\tabular{ll}{
Package: \tab ARSpackage\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-18\cr
License: \tab Wutt \cr
Depends: \tab methods\cr
Collate: \tab 'adapt_reject.r', 'ARS_methods.r' \cr
}
}
\author{
J. Bladen, L. Felberg, H.W. Tsao, S. Tu
Maintainer: L. Felberg <lfelberg@berkeley.edu>
}
\references{
Gilks, Wild, 1992. \url{http://faculty.chicagobooth.edu/hedibert.lopes/teaching/ccis2010/1992GilksWild.pdf}.
}
\keyword{ package, rejection sampling }
\seealso{
\url{https://github.com/paciorek/stat243-fall-2013/tree/master/project}
}
\examples{
~~ simple examples of the most important functions ~~
}
