\name{Cadapt_reject_sample}
\alias{Cadapt_reject_sample}
\title{The adapt_reject class}
\description{
  This class contains all the methods used to perform an AR
  sampling.
}
\section{Slots}{
  \describe{ \item{\code{n}:}{Variable of class
  \code{"numeric"}, n, containing the number of points to
  sample, taken as user input.} \item{\code{f_x}:}{Function
  of class \code{"function"}, containing the f(x) to sample
  from, taken as user input.}
  \item{\code{bounds}:}{Variable of class \code{"numeric"},
  n, containing the bounds of the function, taken as user
  input.} \item{\code{output}:}{Variable of class
  \code{"vector"}, containing sampled points to return to
  user.} \item{\code{h_at_x}:}{Variable of class
  \code{"vector"}, containing computed log(f(x)) values at
  all x values} \item{\code{hprime_at_x}:}{Variable of
  class \code{"vector"}, containing computed derivative of
  log(f(x)) values at all x values}
  \item{\code{z}:}{Variable of class \code{"vector"},
  containing abscissae of upper bound function.}
  \item{\code{samples}:}{Variable of class \code{"vector"},
  containing random numbers generated by s(x) and unif.}
  \item{\code{x}:}{Variable of class \code{"vector"},
  containing x values used in ARS.}
  \item{\code{weights}:}{Variable of class \code{"vector"},
  containing sampled points to return to user.}
  \item{\code{output}:}{Variable of class \code{"numeric"},
  containing sampled points to return to user.}
  \item{\code{mat_sorted}:}{Variable of class
  \code{"matrix"}, containing x values, their corresponding
  h and h prime values, sorted by increasing x.}
  \item{\code{guess_of_mode}:}{Variable of class
  \code{"numeric"}, containing an optional user input guess
  of the mode of the distribution, should be within 200 of
  actual mode.} }
}

