\name{initialize}
\alias{initialize}
\title{Cadapt_reject_sample initialization: method to intialize the ARS class for sampling}
\arguments{
  \item{object}{\code{\linkS4class{Cadapt_reject_sample}}
  object}

  \item{n}{\code{numeric} determining the number of samples
  to obtain}

  \item{f_x}{\code{function} for distribution to sample
  from}

  \item{bounds}{\code{vector} of distribution bounds}
}
\description{
  Cadapt_reject_sample initialization: method to intialize
  the ARS class for sampling
}

