\name{validity_ars}
\alias{validity_ars}
\title{Validity checks for S4 \code{adapt_reject_sample} object: want to ensure at creation that the number of samples desired is a positive integer}
\usage{
  validity_ars(object)
}
\arguments{
  \item{object}{An \code{adapt_reject_sample} object}
}
\description{
  Validity checks for S4 \code{adapt_reject_sample} object:
  want to ensure at creation that the number of samples
  desired is a positive integer
}

